requirements.txt
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pytest>=7.0.0
python-docx>=0.8.11
anthropic>=0.25.0
python-dotenv>=1.0.0
